feat(documents)!: update UploadFile component and route allowing upload from /documents with metadata, temporarily preserve old scripts
